refactor(restir): split the domain-neutral reservoir core out of ReSTIR DI
#1169 asks for Reservoir.glsl to be shared "where the shape is genuinely
shared" and forked where it is not, and the test for that is one question: does
a piece refer to what the sample IS? Streaming RIS does not care whether the
survivor is a point on an emitter or a one-bounce hit vertex, and neither does
the normaliser, the balance heuristic, the M cap, the GEOMETRIC CORE of the
reconnection Jacobian, or the arithmetic that gets an integer through a float
lane.

So those move to ReservoirCore.h / include/ReservoirCore.glsl, and what stays
in ReservoirDI.h / Reservoir.glsl is what a LIGHT SAMPLE makes true: the sample
kinds, the delta-light arm of the shift, and the DI plane layout. The geometric
core is shared because the derivation - one vertex held fixed, a solid-angle
density re-expressed at a moved shading point - never looks at what the vertex
is; the degenerate arms are forked because that is exactly where the two
domains differ.

GLSL has no templates, so the four state operations that need the reservoir and
sample TYPES come from one macro. That keeps the selection test - the line that
is easy to get wrong and impossible to spot wrong - in one place, which is the
whole point.

Two further extractions, for the same reason #1140 extracted the light
densities: a tier validated against an oracle must not carry its own copy of
what the oracle sees.

- include/RayTracedSurfaceHit.glsl takes the material-texture table, the glTF
  MASK alpha test, the candidate-solid rule, the ray flags, the tangent frame,
  the triangle fetch, the closest-hit resolve and the segment occlusion query
  out of GpuPathTracer.glsl, together with the load-bearing include order for
  LightSampling.glsl and RayTracingAlphaTest.glsl.
- include/GBufferRaySurface.glsl takes the G-Buffer shading-point
  reconstruction out of ReSTIRDICommon.glsl, so both ReSTIR tiers start their
  rays from the same point.

No behaviour moved, and that is checked rather than asserted. Compiling every
touched shader with glslc before and after and comparing SPIR-V opcode
histograms: GpuPathTracer, ReSTIR_DI_TemporalReuse and ReSTIR_DI_Resolve are
byte-for-byte identical apart from the moved functions' debug names;
ReSTIR_DI_InitialSample and ReSTIR_DI_SpatialReuse differ only by the two
extracted helpers appearing as their own functions instead of inline (+1
OpFunction, +1 OpFunctionCall and the parameters that implies). No arithmetic,
conversion, bitwise or control-flow opcode count changed in any of them.

ReSTIRDIContractTest now scans the core file too, and gains a float scanner so
the encoding constants that moved - the kind mask, the octahedral scale and
stride, the exact-integer limit, the minimum shift cosine - keep being checked
rather than quietly stopping being checked, which looks the same from a green
run.
