+++ OloEditor/assets/shaders/include/ReSTIRDICommon.glsl
@@ -33,78 +33,25 @@
 #include "ReSTIRDIParams.glsl"
 
 // ---------------------------------------------------------------------------
-// The surface
+// The surface — include/GBufferRaySurface.glsl's, under DI's names
 // ---------------------------------------------------------------------------
+//
+// The G-Buffer reconstruction moved there for issue #1169, so ReSTIR DI and
+// ReSTIR GI reconstruct the SAME shading point from the same texels. Both tiers
+// trace rays from it and both convert densities through it, so a half-texel
+// difference between two copies would make one tier's rays start somewhere the
+// other's do not — which reads as "the two tiers disagree slightly about
+// contact shadowing" rather than as a reconstruction bug.
+//
+// The names below are ALIASES: no DI draw changed, and the compiled SPIR-V is
+// unmoved.
+#include "GBufferRaySurface.glsl"
 
-// What a pixel's G-Buffer says, in the render-relative frame the TLAS is built
-// in (issue #429). Reading absolute world positions here would displace every
-// ray by exactly the render origin — zero near the world origin, which is where
-// every test and benchmark scene sits, and a 1024 m miss the moment the origin
-// grid snaps.
-struct OloReSTIRSurface
-{
-    bool Valid;
-    vec3 Position;
-    vec3 GeometricNormal;
-    vec3 ShadingNormal;
-    vec3 Albedo;
-    float Metallic;
-    float Roughness;
-    float ViewDepth;
-    int PbrModel;
-};
-
-// Octahedral decode — matches octEncodeGB() in PBR_GBuffer.glsl.
-vec3 OloReSTIROctDecode(vec2 e)
-{
-    vec3 n = vec3(e, 1.0 - abs(e.x) - abs(e.y));
-    if (n.z < 0.0)
-        n.xy = (1.0 - abs(n.yx)) * vec2(n.x >= 0.0 ? 1.0 : -1.0, n.y >= 0.0 ? 1.0 : -1.0);
-    return normalize(n);
-}
-
-vec3 OloReSTIRViewPosFromDepth(vec2 uv, float depth)
-{
-    vec4 ndc = vec4(uv * 2.0 - 1.0, depth * 2.0 - 1.0, 1.0);
-    vec4 view = u_InvProjection * ndc;
-    return view.xyz / view.w;
-}
-
-// The G-Buffer gives ONE normal, so the geometric and shading normals are the
-// same vector here. That is stated rather than hidden: it means a normal-mapped
-// surface offsets its own ray origin along the mapped normal, which is the same
-// compromise RayTracedShadow.glsl and every other G-Buffer ray consumer makes,
-// and it is why RayOriginNormalBias exists as a tunable.
-OloReSTIRSurface OloReSTIRLoadSurface(vec2 uv, float depth, vec4 packedNormal, vec4 packedAlbedo, int pbrModel)
-{
-    OloReSTIRSurface s;
-    s.Valid = true;
-    vec3 viewPos = OloReSTIRViewPosFromDepth(uv, depth);
-    s.Position = (u_InvView * vec4(viewPos, 1.0)).xyz;
-    s.ViewDepth = -viewPos.z;
-    s.ShadingNormal = OloReSTIROctDecode(packedNormal.xy);
-    s.GeometricNormal = s.ShadingNormal;
-    s.Roughness = packedNormal.z;
-    s.Albedo = packedAlbedo.rgb;
-    s.Metallic = packedAlbedo.a;
-    s.PbrModel = pbrModel;
-    return s;
-}
-
-OloReSTIRSurface OloReSTIRInvalidSurface()
-{
-    OloReSTIRSurface s;
-    s.Valid = false;
-    s.Position = vec3(0.0);
-    s.GeometricNormal = vec3(0.0, 0.0, 1.0);
-    s.ShadingNormal = vec3(0.0, 0.0, 1.0);
-    s.Albedo = vec3(0.0);
-    s.Metallic = 0.0;
-    s.Roughness = 1.0;
-    s.ViewDepth = 0.0;
-    s.PbrModel = 0;
-    return s;
-}
+#define OloReSTIRSurface OloGBufferSurface
+#define OloReSTIROctDecode OloGBufferOctDecode
+#define OloReSTIRViewPosFromDepth OloGBufferViewPosFromDepth
+#define OloReSTIRLoadSurface OloLoadGBufferSurface
+#define OloReSTIRInvalidSurface OloInvalidGBufferSurface
 
 // ---------------------------------------------------------------------------
 // The target function
